refactor(core): move timer to seperate package
